🚚 Move google-app-check functions to a subdirectory
