+++ src/functions/google-app-check/index.ts
@@ -1,0 +1,1 @@
+export { GoogleAppCheckGenerateToken } from './generate-token.js';
